Fix dom-utils.ts
