+++ client/shared/src/testing/dom-utils.ts
@@ -1,5 +1,4 @@
 // Inspired by React's packages/dom-event-testing-library/domEnvironment.js
-import { Driver } from './driver'
 
 const originalPlatform = window.navigator.platform
 const platformGetter = jest.spyOn(window.navigator, 'platform', 'get')
@@ -26,19 +25,4 @@ export const platform = {
             }
         }
     },
-}
-
-/**
- * Checks that given element has page focus.
- *
- * @param driver - testing page driver
- * @param selector - selector for looking an element that should have focus
- */
-export async function hasFocus(driver: Driver, selector: string): Promise<boolean> {
-    return driver.page.evaluate(selector => {
-        const element = document.querySelector<HTMLElement>(selector)
-        const focusedElement = document.activeElement
-
-        return element === focusedElement
-    }, selector)
 }
